se ha implementado la autenticacion para products

+++ app/Http/Controllers/ProductController.php
@@ -10,6 +10,7 @@ use App\Http\Resources\ProductCollection;
 use App\Http\Resources\ProductResource;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
@@ -18,30 +19,43 @@ class ProductController extends Controller
      */
     public function index(Request $request)
     {
-        $filter = new ProductFilters();
-        $queryItems = $filter->useTransform($request);
+        $user = Auth::user();
+        if ($user) {
+            $tokens = $user->tokens;
+            $hasAccess = $tokens->contains(function ($token) {
+                return isset($token->abilities['products']) && in_array('read', $token->abilities['products']);
+            });
+            if (!$hasAccess) {
+                return response()->json(['error' => 'You are not authorized to view this information'], 403);
+            } else {
+                $filter = new ProductFilters();
+                $queryItems = $filter->useTransform($request);
 
-        // Inicializamos la consulta
-        $productsQuery = Product::query();
+                // Inicializamos la consulta
+                $productsQuery = Product::query();
 
-        // Aplicamos las condiciones de filtro a la consulta
-        foreach ($queryItems as $queryItem) {
-            $column = $queryItem[0]; // Nombre de la columna
-            $operator = $queryItem[1]; // Operador
-            $value = $queryItem[2]; // Valor
+                // Aplicamos las condiciones de filtro a la consulta
+                foreach ($queryItems as $queryItem) {
+                    $column = $queryItem[0]; // Nombre de la columna
+                    $operator = $queryItem[1]; // Operador
+                    $value = $queryItem[2]; // Valor
 
-            // Aplicamos la condición a la consulta
-            $productsQuery->where($column, $operator, $value);
+                    // Aplicamos la condición a la consulta
+                    $productsQuery->where($column, $operator, $value);
+                }
+
+                // Obténemos la cantidad total de los productos que coinciden con los filtros
+                $totalProducts = $productsQuery->count();
+
+                // Si hay más de 10 productos, paginamos; de lo contrario, obtenenemos todos los productos
+                $products = $totalProducts > 10 ? $productsQuery->paginate() : $productsQuery->get();
+
+                // Devuelvemos la colección de productos
+                return new ProductCollection($products);
+            }
+        } else {
+            return response()->json(['code' => 403, 'unauthenticated']);
         }
-
-        // Obténemos la cantidad total de los productos que coinciden con los filtros
-        $totalProducts = $productsQuery->count();
-
-        // Si hay más de 10 productos, paginamos; de lo contrario, obtenenemos todos los productos
-        $products = $totalProducts > 10 ? $productsQuery->paginate() : $productsQuery->get();
-
-        // Devuelvemos la colección de productos
-        return new ProductCollection($products);
     }
 
     /**
@@ -57,15 +71,29 @@ class ProductController extends Controller
      */
     public function store(StoreProductRequest $request)
     {
-        // Verificar si el SKU es único
-        if ($request->sku && !$this->skuUnique($request->sku, null)) {
-            return response()->json(['message' => 'The SKU is already in use by another product.'], 422);
+
+        $user = Auth::user();
+        if ($user) {
+            $tokens = $user->tokens;
+            $hasAccess = $tokens->contains(function ($token) {
+                return isset($token->abilities['products']) && in_array('create', $token->abilities['products']);
+            });
+            if (!$hasAccess) {
+                return response()->json(['error' => 'You are not authorized for this operation'], 403);
+            } else {
+                // Verificar si el SKU es único
+                if ($request->sku && !$this->skuUnique($request->sku, null)) {
+                    return response()->json(['message' => 'The SKU is already in use by another product.'], 422);
+                }
+
+                // Crear el nuevo producto
+                $newProduct = Product::create($request->all());
+
+                return new ProductResource($newProduct);
+            }
+        } else {
+            return response()->json(['code' => 403, 'unauthenticated']);
         }
-    
-        // Crear el nuevo producto
-        $newProduct = Product::create($request->all());
-    
-        return new ProductResource($newProduct);
     }
 
     /**
@@ -73,11 +101,26 @@ class ProductController extends Controller
      */
     public function show($productId)
     {
-        try {
-            $product = Product::findOrFail($productId);
-            return new ProductResource($product);
-        } catch (ModelNotFoundException $exception) {
-            return response()->json(['message' => 'Product not found'], 404);
+
+        $user = Auth::user();
+        if ($user) {
+            $tokens = $user->tokens;
+            $hasAccess = $tokens->contains(function ($token) {
+                return isset($token->abilities['products']) && in_array('read', $token->abilities['products']);
+            });
+            if (!$hasAccess) {
+                return response()->json(['error' => 'You are not authorized to view this information'], 403);
+            } else {
+
+                try {
+                    $product = Product::findOrFail($productId);
+                    return new ProductResource($product);
+                } catch (ModelNotFoundException $exception) {
+                    return response()->json(['message' => 'Product not found'], 404);
+                }
+            }
+        } else {
+            return response()->json(['code' => 403, 'unauthenticated']);
         }
     }
 
@@ -95,24 +138,38 @@ class ProductController extends Controller
      */
     public function update(UpdateProductRequest $request, $productId)
     {
-        try {
-            $product = Product::findOrFail($productId);
-        } catch (ModelNotFoundException $exception) {
-            return response()->json(['message' => 'Product not found'], 404);
+
+        $user = Auth::user();
+        if ($user) {
+            $tokens = $user->tokens;
+            $hasAccess = $tokens->contains(function ($token) {
+                return isset($token->abilities['products']) && in_array('update', $token->abilities['products']);
+            });
+            if (!$hasAccess) {
+                return response()->json(['error' => 'You are not authorized for this operation'], 403);
+            } else {
+                try {
+                    $product = Product::findOrFail($productId);
+                } catch (ModelNotFoundException $exception) {
+                    return response()->json(['message' => 'Product not found'], 404);
+                }
+
+                // Verificar si el SKU es único, excluyendo el producto actual
+                // if ($request->sku && !$this->skuUnique($request->sku, $productId)) {
+                //     return response()->json(['message' => 'The SKU is already in use by another product.'], 422);
+                // }
+
+                // Excluir el campo SKU del array de datos para asegurar que no se modifique
+                $requestData = $request->except('sku');
+
+                // Actualizar el producto
+                $product->update($requestData);
+
+                return new ProductResource($product);
+            }
+        } else {
+            return response()->json(['code' => 403, 'unauthenticated']);
         }
-    
-        // Verificar si el SKU es único, excluyendo el producto actual
-        // if ($request->sku && !$this->skuUnique($request->sku, $productId)) {
-        //     return response()->json(['message' => 'The SKU is already in use by another product.'], 422);
-        // }
-    
-        // Excluir el campo SKU del array de datos para asegurar que no se modifique
-        $requestData = $request->except('sku');
-    
-        // Actualizar el producto
-        $product->update($requestData);
-    
-        return new ProductResource($product);
     }
 
     private function skuUnique($sku, $productId)
@@ -127,13 +184,27 @@ class ProductController extends Controller
      */
     public function destroy($productId)
     {
-        $product = Product::find($productId);
 
-        if ($product) {
-            $product->delete();
-            return response()->json(['message' => 'Product deleted successfully'], 200);
+        $user = Auth::user();
+        if ($user) {
+            $tokens = $user->tokens;
+            $hasAccess = $tokens->contains(function ($token) {
+                return isset($token->abilities['products']) && in_array('delete', $token->abilities['products']);
+            });
+            if (!$hasAccess) {
+                return response()->json(['error' => 'You are not authorized for this operation'], 403);
+            } else {
+                $product = Product::find($productId);
+
+                if ($product) {
+                    $product->delete();
+                    return response()->json(['message' => 'Product deleted successfully'], 200);
+                } else {
+                    return response()->json(['message' => 'Product not found'], 404);
+                }
+            }
         } else {
-            return response()->json(['message' => 'Product not found'], 404);
+            return response()->json(['code' => 403, 'unauthenticated']);
         }
     }
 }
